Shader for text test
Shader for text test, add global classes, GAPI changed

+++ Core/Utilities.cs
@@ -5,6 +5,6 @@ namespace GEngine.Core
     {
         public static float GetPercentByScreen(int pixel, bool isWidth) => (float)pixel / Config.pixelScreenWidth;
         public static Vector2 GetPercentByScreen(Vector2 pixel)
-            => new Vector2((float)pixel.X / Config.pixelScreenWidth, (float)pixel.Y / Config.pixelScreenHeight);
+            => new Vector2(pixel.X / Config.pixelScreenWidth, pixel.Y / Config.pixelScreenHeight);
     }
 }
